split tests and make more verbose
